Updated HMPdistance
Added the source.directly.from.github option: if FALSE (default) requires that the user downloads the taxonomy-GG13-8-and-HMP-reference-sets-L2-L6.rdata and save it to your current directory, it will be loaded automatically by the function. The file contains the reference sets and the GG13.8 taxonomy object, all are needed by the HMPdistance function. If TRUE, the file will be laoded directly from GitHub (requires that **repmis** library is loaded).